i added warnings instead of exceptions in the quench stuff

+++ src/tools/SwitchingFuncs.jl
@@ -1,5 +1,5 @@
-_cos4(τ)::ComplexF64 = abs(real(τ)) <= 1 ? cos(π*τ/2)^4 : 0
-_gauss(τ)::ComplexF64 = exp(-τ^2/2σ) 
+_cos4(τ)  = abs(real(τ)) <= 1 ? cos(π*τ/2)^4 : 0
+_gauss(τ) = exp(-τ^2/2σ) 
 
 χs = Dict("cos4"  => _cos4,
           "gauss" => _gauss,
